refactor: hide ZenCmsLite and build_source_transform from public API
ZenCmsLite is pub(crate) — not yet benchmarked against moxcms on
aarch64 and the SIMD story is incomplete. build_source_transform
removed from ColorManagement trait — the trait is identical to v0.2.3.

finalize_for_output uses moxcms-first ICC-to-ICC path (matching v0.2.3
behavior). build_cms_transform simplified to ICC-bytes-only.

Integration tests using ZenCmsLite moved to unit tests in cms_lite.rs.
