+++ zenpixels-convert/src/lib.rs
@@ -392,7 +392,7 @@ pub(crate) mod negotiate;
 pub mod adapt;
 pub mod cms;
 #[cfg(feature = "zencms-lite")]
-pub mod cms_lite;
+pub(crate) mod cms_lite;
 #[cfg(feature = "cms-moxcms")]
 pub mod cms_moxcms;
 pub mod converter;
@@ -443,7 +443,8 @@ pub use hdr::{
 // Re-export CMS traits, enums, and implementations.
 pub use cms::{ColorManagement, ColorPriority, RenderingIntent, RowTransform};
 #[cfg(feature = "zencms-lite")]
-pub use cms_lite::ZenCmsLite;
+// TODO: pub use cms_lite::ZenCmsLite once benchmarked against moxcms on
+// aarch64 and the SIMD story is complete. See also build_source_transform.
 #[cfg(feature = "cms-moxcms")]
 pub use cms_moxcms::MoxCms;
