+++ zenpixels-convert/src/output.rs
@@ -288,14 +288,11 @@ pub fn finalize_for_output<C: ColorManagement>(
     })
 }
 
-/// Build a CMS transform from the origin's authoritative color metadata.
+/// Build a CMS transform from the origin's color metadata.
 ///
 /// Respects [`ColorAuthority`]: when `Icc`, builds from ICC bytes; when `Cicp`,
-/// builds from CICP codes via [`build_source_transform`](ColorManagement::build_source_transform).
-///
-/// Falls back to the non-authoritative field when the authoritative one is
-/// missing (e.g., `Icc` authority but no ICC → tries CICP). This handles
-/// codec bugs (wrong authority) and incomplete metadata gracefully.
+/// builds from CICP codes via the CMS's `build_transform_from_cicp`. Falls
+/// back to the non-authoritative field when the authoritative one is missing.
 ///
 /// Returns `Ok(None)` when no source profile can be determined.
 fn build_cms_transform<C: ColorManagement>(
@@ -309,60 +306,28 @@ fn build_cms_transform<C: ColorManagement>(
     let Some(ref dst_icc) = metadata.icc else {
         return Ok(None);
     };
-    let dst_source = crate::ColorProfileSource::Icc(dst_icc);
 
-    // Try the authoritative source first, then fall back to the other field.
+    // Try ICC path first (or second, depending on authority).
+    let try_icc = |src_icc: &[u8]| -> Result<Option<_>, At<ConvertError>> {
+        let transform = cms
+            .build_transform_for_format(src_icc, dst_icc, src_format, dst_format)
+            .map_err(|e| whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}"))))?;
+        Ok(Some(transform))
+    };
+
     match origin.color_authority {
         ColorAuthority::Icc => {
             if let Some(ref src_icc) = origin.icc {
-                // Try build_source_transform first (handles PrimariesTransferPair fast path)
-                let src = crate::ColorProfileSource::Icc(src_icc);
-                if let Some(result) =
-                    cms.build_source_transform(src, dst_source.clone(), src_format, dst_format)
-                {
-                    let transform = result.map_err(|e| {
-                        whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}")))
-                    })?;
-                    return Ok(Some(transform));
-                }
-                // Fall back to build_transform_for_format (ICC-only path)
-                let transform = cms
-                    .build_transform_for_format(src_icc, dst_icc, src_format, dst_format)
-                    .map_err(|e| whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}"))))?;
-                return Ok(Some(transform));
+                return try_icc(src_icc);
             }
-            // Fallback: ICC authority but no ICC — try CICP if available.
-            if let Some(cicp) = origin.cicp {
-                let src = crate::ColorProfileSource::Cicp(cicp);
-                if let Some(result) =
-                    cms.build_source_transform(src, dst_source, src_format, dst_format)
-                {
-                    let transform = result.map_err(|e| {
-                        whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}")))
-                    })?;
-                    return Ok(Some(transform));
-                }
-            }
+            // Fallback: ICC authority but no ICC bytes — can't build transform.
             Ok(None)
         }
         ColorAuthority::Cicp => {
-            if let Some(cicp) = origin.cicp {
-                let src = crate::ColorProfileSource::Cicp(cicp);
-                if let Some(result) =
-                    cms.build_source_transform(src, dst_source.clone(), src_format, dst_format)
-                {
-                    let transform = result.map_err(|e| {
-                        whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}")))
-                    })?;
-                    return Ok(Some(transform));
-                }
-            }
-            // Fallback: CICP authority but no CICP — try ICC if available.
+            // CICP authority — but build_transform_from_cicp needs ICC bytes
+            // on the dst side, so we still need ICC. Try src ICC if available.
             if let Some(ref src_icc) = origin.icc {
-                let transform = cms
-                    .build_transform_for_format(src_icc, dst_icc, src_format, dst_format)
-                    .map_err(|e| whereat::at!(ConvertError::CmsError(alloc::format!("{e:?}"))))?;
-                return Ok(Some(transform));
+                return try_icc(src_icc);
             }
             Ok(None)
         }
